优化代码

+++ subscribe/index.sh
@@ -1,15 +1,24 @@
 #!/bin/sh
 cd `dirname $0`
 subs_path='/etc/v2ray/subs'
+is_empty_dir(){ 
+    return `ls -A $1|wc -w`
+}
 
 download_config() {
-    rm -f $subs_path/*.json
+    if is_empty_dir ./conf
+    then
+        echo 'no configfile downloaded'
+    else
+        rm -f $subs_path/*.json
+        mv -f ./conf/*.json $subs_path/
+    fi
 }
 
 use_config() {
     configfile=`ls $subs_path/ | grep $1 | sort  -n | head -n 1`
     if [ ! -n "$configfile" ]; then
-        echo "configfile not exists, use the first configfile"
+        echo "configfile not exists"
     fi
     
     if [ -n "$configfile" ]; then
@@ -24,13 +33,13 @@ if [ "$#" = 0 ]; then
     download_config
     echo "Enter EndPoint:"
     read cfg
-    if [ ! -n "$cfg" ]; then
+    if [ -n "$cfg" ]; then
         use_config $cfg
     fi
 elif [ "$#" = 1 ] && [ "$1" = "-d" ]; then
     download_config
 elif [ "$#" = 1 ] && [ "$1" = "-r" ]; then
-    use_config "*.config"
+    use_config "*.json"
 elif [ "$#" = 1 ] && [ "$1" != "-h" ]; then
     use_config $1
 else
